+++ common/src/KokkosFFT_Extents.hpp
@@ -11,6 +11,7 @@
 #include <numeric>
 #include <Kokkos_Core.hpp>
 #include "KokkosFFT_Asserts.hpp"
+#include "KokkosFFT_CheckConditions.hpp"
 #include "KokkosFFT_Common_Types.hpp"
 #include "KokkosFFT_Mapping.hpp"
 #include "KokkosFFT_Traits.hpp"
