+++ common/src/KokkosFFT_utils.hpp
@@ -19,9 +19,28 @@
 namespace KokkosFFT {
 namespace Impl {
 
-template <typename ScalarType1, typename ScalarType2>
-bool are_aliasing(const ScalarType1* ptr1, const ScalarType2* ptr2) {
-  return (static_cast<const void*>(ptr1) == static_cast<const void*>(ptr2));
+/// \brief Converts an axis in [-rank, rank-1] to [0, rank-1]
+/// \tparam IntType The integer type used for axis
+///
+/// \param[in] axis The axis to be converted
+/// \param[in] rank The rank of the view
+/// \return The converted axis
+/// \throws a runtime_error if axis is out of range
+template <typename IntType>
+IntType convert_negative_axis(IntType axis, std::size_t rank) {
+  static_assert(std::is_integral_v<IntType>,
+                "convert_negative_axis: IntType must be an integral type.");
+
+  const IntType irank = static_cast<IntType>(rank);
+  if constexpr (std::is_signed_v<IntType>) {
+    KOKKOSFFT_THROW_IF(axis < -irank || axis >= irank,
+                       "Axis must be in [-rank, rank-1]");
+
+    return axis < 0 ? irank + axis : axis;
+  } else {
+    KOKKOSFFT_THROW_IF(axis >= irank, "Axis must be in [0, rank-1]");
+    return axis;
+  }
 }
 
 /// \brief Converts axes in [-rank, rank-1] to [0, rank-1]
@@ -55,83 +74,6 @@ auto convert_negative_axes(const std::array<IntType, DIM>& axes,
                  convert_negative_axis);
 
   return non_negative_axes;
-}
-
-template <typename ContainerType, typename ValueType>
-bool is_found(const ContainerType& values, const ValueType value) {
-  using value_type = KokkosFFT::Impl::base_container_value_type<ContainerType>;
-  static_assert(std::is_same_v<value_type, ValueType>,
-                "is_found: Container value type must match ValueType");
-  return std::find(values.begin(), values.end(), value) != values.end();
-}
-
-template <typename ContainerType>
-bool has_duplicate_values(const ContainerType& values) {
-  using value_type = KokkosFFT::Impl::base_container_value_type<ContainerType>;
-  std::set<value_type> set_values(values.begin(), values.end());
-  return set_values.size() < values.size();
-}
-
-template <typename ContainerType, typename IntType>
-bool is_out_of_range_value_included(const ContainerType& values, IntType max) {
-  static_assert(
-      std::is_integral_v<IntType>,
-      "is_out_of_range_value_included: IntType must be an integral type");
-  using value_type = KokkosFFT::Impl::base_container_value_type<ContainerType>;
-  static_assert(std::is_same_v<value_type, IntType>,
-                "is_out_of_range_value_included: Container value type must "
-                "match IntType");
-  if constexpr (std::is_signed_v<value_type>) {
-    KOKKOSFFT_THROW_IF(
-        std::any_of(values.begin(), values.end(),
-                    [](value_type value) { return value < 0; }),
-        "is_out_of_range_value_included: values must be non-negative");
-  }
-  return std::any_of(values.begin(), values.end(),
-                     [max](value_type value) { return value >= max; });
-}
-
-template <
-    typename ViewType, template <typename, std::size_t> class ArrayType,
-    typename IntType, std::size_t DIM = 1,
-    std::enable_if_t<Kokkos::is_view_v<ViewType> && std::is_integral_v<IntType>,
-                     std::nullptr_t> = nullptr>
-bool are_valid_axes(const ViewType& /*view*/,
-                    const ArrayType<IntType, DIM>& axes) {
-  static_assert(Kokkos::is_view_v<ViewType>,
-                "are_valid_axes: ViewType must be a Kokkos::View");
-  static_assert(std::is_integral_v<IntType>,
-                "are_valid_axes: IntType must be an integral type");
-  static_assert(
-      DIM >= 1 && DIM <= ViewType::rank(),
-      "are_valid_axes: the Rank of FFT axes must be between 1 and View rank");
-
-  // Convert the input axes to be in the range of [0, rank-1]
-  // int type is chosen for consistency with the rest of the code
-  // the axes are defined with int type
-  std::array<IntType, DIM> non_negative_axes = {};
-
-  // In case axis is out of range, 'convert_negative_axes' will throw an
-  // runtime_error and we will return false. Without runtime_error, it is
-  // ensured that the 'non_negative_axes' are in the range of [0, rank-1]
-  try {
-    non_negative_axes = convert_negative_axes(axes, ViewType::rank());
-  } catch (std::runtime_error& e) {
-    return false;
-  }
-
-  bool is_valid = !KokkosFFT::Impl::has_duplicate_values(non_negative_axes);
-  return is_valid;
-}
-
-template <typename ContainerType, typename ValueType>
-std::size_t get_index(const ContainerType& values, const ValueType value) {
-  using value_type = KokkosFFT::Impl::base_container_value_type<ContainerType>;
-  static_assert(std::is_same_v<value_type, ValueType>,
-                "get_index: Container value type must match ValueType");
-  auto it = std::find(values.begin(), values.end(), value);
-  KOKKOSFFT_THROW_IF(it == values.end(), "value is not included in values");
-  return it - values.begin();
 }
 
 template <typename IntType, std::size_t N, IntType start>
