+++ fft/src/KokkosFFT_Transform.hpp
@@ -10,6 +10,7 @@
 #include <Kokkos_Profiling_ScopedRegion.hpp>
 
 #include "KokkosFFT_Asserts.hpp"
+#include "KokkosFFT_CheckConditions.hpp"
 #include "KokkosFFT_Extents.hpp"
 #include "KokkosFFT_Layout.hpp"
 #include "KokkosFFT_Plans.hpp"
